+++ src/data/index.js
@@ -1,8 +1,6 @@
 export {
   fetchCourse,
   fetchSequence,
-  checkBlockCompletion,
-  saveSequencePosition,
 } from './thunks';
 
 export { reducer } from './slice';
